src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
PyMuPDF==1.24.3
Motor==3.1.1
pymongo==4.3.3
pydantic-mongo==2.3.0
qdrant-client==1.10.1
openai==1.77.0
Jinja2==3.1.5
pdfkit==1.0.0
pydantic[email]
passlib[bcrypt]
python-jose
cryptography
typing-extensions
weasyprint
pyppeteer
cohere
pycountry
secure-smtplib



